Fix pause button bug, allegedly
Untested fix to the pause button bug, but in theory should work. Just blocks the user from "unpausing" if there's no wave running since unpausing is handled by "sending wave" which usually just resumes the current wave or sends the next one.

Added an if to the else statement so that repeated presses of the pause button out of wave would not continually decrease the volume of the music.