Fixing tower selection UI bug
Currently everytime a tower is clicked, a dispatch of type "towerClick" is sent with the archtype 

const type = ["Tower_Cannon1", "Tower_Cannon2", "Tower_Cannon3"].includes(archtype) ? "TowerCannon" : "TowerLaser";

If the tower is on the fourth upgrade, it will always return TowerLaser. I'm proposing we update the list to include Tower_Cannon4 to keep it simple. If we desire to add new archtypes in the future, this code will need to be refactored.